CLR wave encoder decoder now only float64 and int64

+++ benchmarks/wave_bench.py
@@ -41,7 +41,7 @@ def main() -> None:
     top_p = 0.95
     use_tqdm = True
     device = "cuda" if torch.cuda.is_available() else "cpu"
-    idx = torch.arange(1_000_000)
+    idx = torch.arange(1_000_000, dtype=torch.float64)
     inp = pd.DataFrame(
         {
             "sin": torch.sin(idx),
@@ -49,7 +49,7 @@ def main() -> None:
         }
     )
     domain_of_definition = get_domain_of_definition(
-        torch.from_numpy(inp.to_numpy()), order_of_derivative=1
+        torch.from_numpy(inp.to_numpy(dtype=np.float64)), order_of_derivative=1
     )
     n = 3
     val_loss_array = torch.zeros((n,))
